reorganize, pushing down def_dialect and friends to rtx_core
