+++ rtx_package/Cargo.toml
@@ -17,7 +17,6 @@ lazy_static = "1.0"
 regex = "1.7.1"
 log = "0.4"
 libxml = "0.3.0"
-unidecode = "0.3.0"
 chrono = "0.4"
 rtx_codegen = {path = "../rtx_codegen"}
 rtx_core = {path = "../rtx_core"}
